move murmur_hash3 to snapshots/common
